Struct deployer::config::ConfigApp

pub struct ConfigApp<'a> {
    // some fields omitted
}

Methods

impl<'a> ConfigApp<'a>

fn confirm_secret(&self, provided: &str) -> bool

Test a provided secret against the configuration. Looks for the application specific secret first before falling back to the default_secret if that's defined. If there is a mismatch, or there are no secrets defined, returns false.

fn playbook(&self, name: &'a String) -> Option<&'a str>

Get the full path of a playbook from the configuration.

fn default_playbook(&self) -> Option<&str>

Get the default playbook for the application if it exists.

fn default_host(&self) -> &str

Get the default host for the app. If there isn't an explicit default set in the configuration, returns "localhost".