Struct deployer::git::GitRepo

pub struct GitRepo {
    pub remote_path: String,
    pub local_path: String,
    pub branch: String,
}

Fields

remote_path

Remote path to the repository. This can be a filesystem path if the file:// protocol is used.

local_path

Local path of where to clone the repository.

branch

Branch to check out. We require this so we can create the smallest checkout possible.

Methods

impl GitRepo

fn fetch(&self) -> Result<Output, CommandError>

If a repo exists, fetch it. If it doesn't, clone it

This is currently very dumb in the sense that it only checks if a directory exists, not whether it's the git repo represented by self or even whether it's a git repository at all.

This is the equivalent of doing:

(test -d <local_path> && cd <local_path> && git fetch) || \
git clone --depth=1 --single-branch -b <branch> <remote_path> <local_path>

fn checkout(&self, sha: &str) -> Result<Output, CommandError>

Check out a specific sha. The equivalent of doing git checkout {sha}.