Struct deployer::message::SimpleMessage

pub struct SimpleMessage {
    pub prefix: Option<String>,
    pub branch: String,
    pub remote: String,
    pub repository_name: String,
}

Fields

prefix

The prefix to differentiate this deployment from another with possibly the same name.

branch

Branch to check out.

remote

Remote path to the repository.

repository_name

Name of the repository. Used to construct the local path where the clone will be stored

Methods

impl SimpleMessage

fn from(json: &str) -> Result<SimpleMessage, &'static str>