Module deployer::task_manager

A task manager for parallel queues that process tasks serially.

When a new queue is created the task manager will spawn a worker thread that immediately blocks. When add_task() is called with the key for that queue the manager will add the task to the queue and fire a signal to the worker thread for that queue that there's a new task waiting. The worker thread then pops a task off the queue and calls its run method. Once there are no more tasks in the queue the worker thread will go back to sleep.

See docs for the TaskManager struct for more usage examples.

Limitations

Task queues can grow infinitely. We should have optional a way to cap the queue size and any tasks that come in after that limit push off the oldest task.

Examples

Waiting for tasks to finish

use deployer::task_manager::{TaskManager, Runnable};
use std::thread;

struct Task { msg: &'static str, delay: u32 };
impl Runnable for Task {
    fn run(&mut self) {
        thread::sleep_ms(self.delay);
        println!("{}", self.msg);
    }
}

let mut task_manager = TaskManager::new();

// This will cause "a", "b", "c" and "1", "2", "3" to print in
// order though letters and numbers will be intermingled because the
// "letters" and "numbers" queues process in parallel.
let last_letter_task = {
    let queue = task_manager.ensure_queue(String::from("letters"));
    task_manager.add_task(&queue, Task {msg: "a", delay: 500});
    task_manager.add_task(&queue, Task {msg: "b", delay: 100});
    task_manager.add_task(&queue, Task {msg: "c", delay: 200})
};

let last_number_task = {
    let queue = task_manager.ensure_queue(String::from("numbers"));
    task_manager.add_task(&queue, Task {msg: "1", delay: 200});
    task_manager.add_task(&queue, Task {msg: "2", delay: 100});
    task_manager.add_task(&queue, Task {msg: "3", delay: 500})
};

last_number_task.unwrap().recv();
last_letter_task.unwrap().recv();

Getting results of a task

use deployer::task_manager::{TaskManager, Runnable};

struct LongRunningTask {
    result: Option<u64>,
}
impl LongRunningTask {
    fn new() -> LongRunningTask {
        LongRunningTask { result: None }
    }
}
impl Runnable for LongRunningTask {
    fn run(&mut self) {
        do_some_hard_work();
        self.result = Some(42);
    }
}

let mut task_manager = TaskManager::new();

let key = task_manager.ensure_queue(String::from("q"));

let task_rx = task_manager.add_task(&key, LongRunningTask::new()).unwrap();
let task = task_rx.recv().unwrap();
assert_eq!(task.result, Some(42));

Graceful shutdowns

let (shutdown_tx, shutdown_rx) = channel();
let task_manager = Arc::new(Mutex::new(TaskManager::new_with_lock(shutdown_tx)));

let shared_manager = task_manager.clone();
event_handler("add_task", move || {
    let mut locked_manager = shared_manager.lock().unwrap();
    let queue = locked_manager.ensure_queue(String::from("q"));
    locked_manager.add_task(&queue, ImportantTask::new()).unwrap();
});

let shared_manager = task_manager.clone();
event_handler("shutdown", move || {
    let mut locked_manager = shared_manager.lock().unwrap();

    // `shutdown()` lets each worker finish a final task if it's already
    // working on one and once each worker is done it sends a message down
    // the lock channel if one exists.
    locked_manager.shutdown();
});

// This blocks until a call to `shutdown()` is completed.
shutdown_rx.recv().unwrap();
println!("task manager done");

Structs

QueueKey

TaskManager

Enums

Error

Traits

Runnable
Types that are able to be added to a TaskManager queue.