Struct deployer::task_manager::TaskManager

pub struct TaskManager<T> where T: 'static + Runnable + Send {
    // some fields omitted
}

Methods

impl<'a, T> TaskManager<T> where T: 'static + Runnable + Send

fn new() -> TaskManager<T>

Create a new TaskManager

fn new_with_lock(lock: Sender<()>) -> TaskManager<T>

Create a new TaskManager that takes a shutdown receiver which can be used to block a thread until shutdown() is called.

fn add_task(&mut self, queue_key: &QueueKey, task: T) -> Result<Receiver<T>, Error>

Add a task to a queue. When the task is complete it will be sent back over the returned Receiver.

Failures

QueueMissing: Could not find the queue. This will only happen if an add_task() call happens before an ensure_queue() call for that queue.

Shutdown: Manager is not accepting new tasks at the moment. This will happen if an add_task() call happens after a shutdown() but before a restart().

fn shutdown(&mut self)

Signal worker threads to shut down after any active jobs and once they are all done send a signal down the shutdown lock channel. Trying to add jobs after calling shutdown() but before calling restart() will result in an Error::Shutdown.

Examples

let (shutdown_tx, shutdown_rx) = channel();
let tasks = Arc::new(Mutex::new(TaskManager::new_with_lock(shutdown_tx)));

let shared_tasks = tasks.clone();
thread::spawn(move || {
    let mut tasks = shared_tasks.lock().unwrap();
    let queue = tasks.ensure_queue(String::from("$"));
    tasks.add_task(&queue, important_task1);
});

// Wait 100ms then signal a shutdown for task manager.
let shared_tasks = tasks.clone();
thread::spawn(move || {
    thread::sleep_ms(100);
    let mut tasks = shared_tasks.lock().unwrap();
    tasks.shutdown();
});

// Wait 150ms then try to add a task. This will fail because the task
// manager will have received a `shutdown()` call by this point.
let shared_tasks = tasks.clone();
thread::spawn(move || {
    thread::sleep_ms(150);
    let mut tasks = shared_tasks.lock().unwrap();
    let queue = tasks.ensure_queue(String::from("$"));
    match tasks.add_task(&queue, important_task2) {
         Err(Error::Shutdown) => println!("queue was shut down"),
         // ...
    };
});

// Block thread until shutdown is complete.
shutdown_rx.recv();

println!("all workers stopped");

fn restart(&mut self)

Restart all queue workers and remove stopped flag.

fn ensure_queue(&mut self, queue_key: String) -> QueueKey

Create a queue only if one doesn't already exist with that key. Returns the QueueKey for that queue.